simplify ci, add 3.14
The venv isn't needed since we're already running
in an ephermal VM. And removing the redundant
default flags from `build` as recommended allow it
to build the wheel from the sdist, which is a good
sanity check to confirm that the sdist contains
everything the wheel needs.

See https://build.pypa.io/en/stable/index.html#python--m-build.